# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sankara Eye Hospital Vision 2020 Conference 2026 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /participants/{participantId}/assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      summary: List assignments for a participant
      parameters:
        - name: participantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
    post:
      operationId: createAssignment
      tags: [assignments]
      summary: Add an assignment to a participant
      parameters:
        - name: participantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /assignments/{id}:
    patch:
      operationId: updateAssignment
      tags: [assignments]
      summary: Update an assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
    delete:
      operationId: deleteAssignment
      tags: [assignments]
      summary: Delete an assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # FILES
...
components:
  schemas:
...
    Assignment:
      type: object
      required: [id, participantId, role, track]
      properties:
        id:
          type: integer
        participantId:
          type: integer
        role:
          type: string
          enum: [Speaker, Presenter, Poster, Panelist, Moderator, Judge, Chair, CoChair]
        track:
          type: string
        sessionName:
          type: ["string", "null"]
        hall:
          type: ["string", "null"]
        date:
          type: ["string", "null"]
        time:
          type: ["string", "null"]
        presentationTitle:
          type: ["string", "null"]
        fileId:
          type: ["integer", "null"]
        uploadedFile:
          $ref: "#/components/schemas/UploadedFile"
          nullable: true
    AssignmentInput:
      type: object
      required: [role, track]
      properties:
        role:
          type: string
          enum: [Speaker, Presenter, Poster, Panelist, Moderator, Judge, Chair, CoChair]
        track:
          type: string
        sessionName:
          type: string
        hall:
          type: string
        date:
          type: string
        time:
          type: string
        presentationTitle:
          type: string
    AssignmentUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [Speaker, Presenter, Poster, Panelist, Moderator, Judge, Chair, CoChair]
        track:
          type: string
        sessionName:
          type: string
        hall:
          type: string
        date:
          type: string
        time:
          type: string
        presentationTitle:
          type: string

    # FILE SCHEMAS
    UploadedFile:
      type: object
      required: [id, assignmentId, filename, originalName, fileType, version, uploadedAt]
      properties:
        id:
          type: integer
        assignmentId:
          type: integer
        filename:
          type: string
        originalName:
          type: string
        fileType:
          type: string
          enum: [pptx, jpg]
        version:
          type: integer
        size:
          type: ["integer", "null"]
        uploadedAt:
          type: string
          format: date-time

    # FOOD SCHEMAS
...